+++ platform/jvm/capture/src/test/kotlin/io/bitdrift/capture/fakes/FakeMemoryMetricsProvider.kt
@@ -44,6 +44,8 @@ class FakeMemoryMetricsProvider : IMemoryMetricsProvider {
             mapOf(
                 "_jvm_used_kb" to "50",
                 "_jvm_total_kb" to "100",
+                "_jvm_max_kb" to "100",
+                "_jvm_used_percent" to "50",
                 "_native_used_kb" to "200",
                 "_native_total_kb" to "500",
                 "_memory_class" to "1",