+++ platform/jvm/common/src/main/kotlin/io/bitdrift/capture/common/Runtime.kt
@@ -29,11 +29,6 @@ sealed class RuntimeFeature(
     data object APP_UPDATE_EVENTS : RuntimeFeature("client_feature.android.application_update_reporting", defaultValue = true)
 
     /**
-     * Whether memory pressure monitoring is enabled.
-     */
-    data object APP_MEMORY_PRESSURE : RuntimeFeature("client_feature.android.memory_pressure_reporting", defaultValue = true)
-
-    /**
      * Whether device state monitoring is enabled.
      */
     data object DEVICE_STATE_EVENTS : RuntimeFeature("client_features.android.device_lifecycle_reporting", defaultValue = true)
@@ -115,6 +110,11 @@ sealed class RuntimeConfig(
      * ANR Frame: >= ANR_FRAME_THRESHOLD_MS
      */
     data object ANR_FRAME_THRESHOLD_MS : RuntimeConfig("client_feature.android.anr_frame.threshold_ms", 5000)
+
+    /**
+     * The App low memory percentage configured threshold where will start reporting is_memory_low event as true. (Default threshold is 90%)
+     */
+    data object APP_LOW_MEMORY_PERCENT_THRESHOLD : RuntimeConfig("client_feature.android.app_low_memory_percent_threshold", 90)
 }
 
 /**